[QPID] Tests for qpid-jms-client and messaging components
