+++ integration-tests/pom.xml
@@ -150,8 +150,8 @@
         <module>lumberjack</module>
         <module>mail</module>
         <module>master</module>
-        <module>messaging</module>
-        <module>messaging-qpid-client</module>
+        <module>messaging-artemis-jms-client</module>
+        <module>messaging-qpid-jms-client</module>
         <module>micrometer</module>
         <module>microprofile</module>
         <module>minio</module>
