+++ tooling/scripts/test-categories.yaml
@@ -134,7 +134,8 @@ messaging-networking1:
   - amqp
   - kafka-sasl-ssl
   - kafka-ssl
-  - messaging
+  - messaging-artemis-jms-client
+  - messaging-qpid-jms-client
   - nats
   - splunk
   - spring-rabbitmq
